+++ src/Oro/Bundle/SalesBundle/Tests/Functional/ImportExport/Strategy/B2bConfigurableAddOrReplaceStrategyTest.php
@@ -51,7 +51,7 @@ class B2bConfigurableAddOrReplaceStrategyTest extends WebTestCase
 
         $this->strategy = new B2bConfigurableAddOrReplaceStrategy(
             $container->get('event_dispatcher'),
-            $container->get('oro_importexport.strategy.import.helper'),
+            $container->get('oro_importexport.strategy.configurable_import_strategy_helper'),
             $container->get('oro_entity.helper.field_helper'),
             $container->get('oro_importexport.field.database_helper'),
             $container->get('oro_entity.entity_class_name_provider'),